Improve item search view
Both pages now allow search and tick items.
