+++ app/src/debug/res/menu/activity_main_drawer.xml
@@ -15,7 +15,7 @@
             android:visible="false"/>
         <item
             android:id="@+id/nav_add_list"
-            android:icon="@drawable/ic_add_shopping_cart_black_24dp"
+            android:icon="@drawable/ic_playlist_add_black_24dp"
             android:title="@string/menu_new_list"/>
     </group>
