+++ app/src/main/res/menu/add_item.xml
@@ -6,7 +6,7 @@
         android:id="@+id/action_add_item"
         android:icon="@drawable/ic_add_shopping_cart_white_24dp"
         android:menuCategory="secondary"
-        android:orderInCategory="100"
+        android:orderInCategory="50"
         android:title="@string/add_item"
         app:showAsAction="ifRoom"/>
